TR2-16 refactor code
