+++ src/main/java/rueppellii/backend2/tribes/troops/WarriorTroop.java
@@ -1,4 +1,4 @@
-package rueppellii.backend2.tribes.kingdom.troops;
+package rueppellii.backend2.tribes.troops;
 
 import javax.persistence.Entity;
 import java.sql.Timestamp;
